start off by making the external_interface module in core its own lib
