+++ libs/core/Cargo.toml
@@ -24,7 +24,6 @@ chrono = "0.4.15"
 diffy = "0.2.0"
 fuzzy-matcher = "0.3.7"
 image = "0.24.3"
-jni = { version = "0.13.1", default-features = false }
 raqote = { version = "0.8.0", default-features = false }
 reqwest = { version = "0.11.1", default-features = false, features = ["blocking", "json"] }
 serde = { version = "1.0", features = ["derive"] }
@@ -39,7 +38,6 @@ tracing = "0.1.5"
 tracing-subscriber = "0.3.9"
 tracing-appender = "0.2"
 hmdb = "0.2.2"
-lazy_static = "1.4.0"
 sublime_fuzzy = "0.7.0"
 crossbeam = "0.8.1"
 lockbook-shared = { path = "libs/shared" }
