--- libs/core/src/external_interface/mod.rs
@@ -1,4 +1,0 @@
-pub mod c_interface;
-pub mod java_interface;
-pub mod json_interface;
-pub mod static_state;
