+++ libs/core/src/lib.rs
@@ -7,8 +7,6 @@ pub mod model;
 pub mod repo;
 pub mod service;
 
-mod external_interface;
-
 pub use base64;
 pub use basic_human_duration::ChronoHumanDuration;
 pub use chrono::Duration;
@@ -54,8 +52,6 @@ use lockbook_shared::api::{
     AccountInfo, AdminFileInfoResponse, AdminValidateAccount, AdminValidateServer,
 };
 use lockbook_shared::clock;
-use serde_json::{json, value::Value};
-use strum::IntoEnumIterator;
 
 use crate::model::errors::Error::UiError;
 use crate::repo::CoreDb;
@@ -475,46 +471,6 @@ impl<Client: Requester> CoreLib<Client> {
 pub fn get_code_version() -> &'static str {
     env!("CARGO_PKG_VERSION")
 }
-
-// This basically generates a function called `get_all_error_variants`,
-// which will produce a big json dict of { "Error": ["Values"] }.
-// Clients can consume this and attempt deserializing each array of errors to see
-// if they are handling all cases
-macro_rules! impl_get_variants {
-    ( $( $name:ty,)* ) => {
-        fn get_all_error_variants() -> Value {
-            json!({
-                $(stringify!($name): <$name>::iter().collect::<Vec<_>>(),)*
-            })
-        }
-    };
-}
-
-// All new errors must be placed in here!
-impl_get_variants!(
-    CreateAccountError,
-    ImportError,
-    AccountExportError,
-    GetAccountError,
-    CreateFileAtPathError,
-    WriteToDocumentError,
-    CreateFileError,
-    GetRootError,
-    GetFileByIdError,
-    GetFileByPathError,
-    FileDeleteError,
-    ReadDocumentError,
-    RenameFileError,
-    MoveFileError,
-    SyncAllError,
-    CalculateWorkError,
-    GetUsageError,
-    GetDrawingError,
-    SaveDrawingError,
-    ExportDrawingError,
-    ExportDrawingToDiskError,
-    SaveDocumentToDiskError,
-);
 
 pub static DEFAULT_API_LOCATION: &str = "https://api.prod.lockbook.net";
 pub static CORE_CODE_VERSION: &str = env!("CARGO_PKG_VERSION");
