+++ libs/core_external_interface/src/lib.rs
@@ -1,0 +1,51 @@
+#[macro_use]
+extern crate tracing;
+
+pub mod c_interface;
+pub mod java_interface;
+pub mod json_interface;
+pub mod static_state;
+
+use lockbook_core::*;
+use serde_json::{json, value::Value};
+use strum::IntoEnumIterator;
+
+// This basically generates a function called `get_all_error_variants`,
+// which will produce a big json dict of { "Error": ["Values"] }.
+// Clients can consume this and attempt deserializing each array of errors to see
+// if they are handling all cases
+macro_rules! impl_get_variants {
+    ( $( $name:ty,)* ) => {
+        fn get_all_error_variants() -> Value {
+            json!({
+                $(stringify!($name): <$name>::iter().collect::<Vec<_>>(),)*
+            })
+        }
+    };
+}
+
+// All new errors must be placed in here!
+impl_get_variants!(
+    CreateAccountError,
+    ImportError,
+    AccountExportError,
+    GetAccountError,
+    CreateFileAtPathError,
+    WriteToDocumentError,
+    CreateFileError,
+    GetRootError,
+    GetFileByIdError,
+    GetFileByPathError,
+    FileDeleteError,
+    ReadDocumentError,
+    RenameFileError,
+    MoveFileError,
+    SyncAllError,
+    CalculateWorkError,
+    GetUsageError,
+    GetDrawingError,
+    SaveDrawingError,
+    ExportDrawingError,
+    ExportDrawingToDiskError,
+    SaveDocumentToDiskError,
+);
